feat: RT100747 ignore inline styling hint
No longer check for inline styling hints like <nobr> elements, 'nowrap'
attributes or inline style attributes setting the 'whitespace' property
to optimise the process of adding hyphens to a tree. This introduces a
suboptimal situation, but does not break the end-user experience since
it only adds soft hyphens to element that don't use them as far as the
parsed tree is concerned. But this does allow to override the default
behavior of the tree with external styles that could take advantage of
the available soft hyphens.